+++ src/app/_components/productDetail/ProductDescription.tsx
@@ -1,0 +1,9 @@
+const ProductDescription = () => {
+  return (
+    <section className="flex items-center w-full max-w-6xl gap-4 p-4">
+      <p>작성자가 작성한 설명</p>
+    </section>
+  );
+};
+
+export default ProductDescription;
