+++ src/app/_components/productDetail/ProductNotice.tsx
@@ -1,0 +1,9 @@
+const ProductNotice = () => {
+  return (
+    <section className="flex items-center w-full max-w-6xl gap-4 p-4">
+      <p>주의사항 안내</p>
+    </section>
+  );
+};
+
+export default ProductNotice;
